MODIFY: route is now secured and better res from the api

+++ src/app/api/expire-session/route.ts
@@ -1,7 +1,19 @@
 import { deleteExpiredSessions } from "@/util/session";
 
-export async function GET() {
-  await deleteExpiredSessions();
+export async function GET(request: Request) {
+  const authHeader = request.headers.get("authorization");
+  if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
+    return new Response("Unauthorized", {
+      status: 401,
+    });
+  }
 
-  return new Response("Expired sessions removed from database");
+  const res = await deleteExpiredSessions();
+
+  return new Response(res, {
+    status: 200,
+    headers: {
+      "Content-Type": "text/plain",
+    },
+  });
 }
